+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.106.0
 * Added option WithReaderSupportSplitMergePartitions for topic manage support of split-merge partitions on client side (enabled by default).
 * Allowed overflow queue limit for one goroutine at time for topic writer
 * Removed delay before send commit in sync mode of a topic reader